refactor(backend): types in seperate file
