+++ Backend/lib/types.ts
@@ -1,0 +1,7 @@
+export type UserType = {
+    id?: number;
+    firstName: string;
+    lastName: string;
+    email: string;
+    password: string;
+};
